requirements.txt
Mastodon.py
bs4
pyyaml
langid
torch
transformers